+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Date format: DD/MM/YYYY
   - `TextButton`
 - `PaneItem`s' `build` method is now overridable. You can know customize how the items in `NavigationView` should look like by overriding the method.
 
+## [2.0.3] - [28/05/2021]
+
+- Correctly apply items positions to pane indicators, regardless of external factors, such as navigation view app bar ([#41](https://github.com/bdlukaa/fluent_ui/issues/41))
+- Improved `NavigationIndicator`s performance
+
 ## [2.0.2] - [23/05/2021]
 
 - **BREAKING CHANGES:** Reworked the theme api ([#39](https://github.com/bdlukaa/fluent_ui/pull/39)):